Keep reprocess_error_forms, delete other one
Move the necessary functionality into reprocess_error_forms
